Rename test helpers module in mqtt
