+++ tests/components/mqtt/common.py
