+++ tests/components/mqtt/test_binary_sensor.py
@@ -23,7 +23,7 @@ from homeassistant.core import HomeAssistant, State, callback
 from homeassistant.helpers.typing import ConfigType
 from homeassistant.util import dt as dt_util
 
-from .test_common import (
+from .common import (
     help_custom_config,
     help_test_availability_when_connection_lost,
     help_test_availability_without_topic,
