+++ tests/components/mqtt/test_light.py
@@ -210,7 +210,7 @@ from homeassistant.components.mqtt.models import PublishPayloadType
 from homeassistant.const import ATTR_ASSUMED_STATE, STATE_OFF, STATE_ON, STATE_UNKNOWN
 from homeassistant.core import HomeAssistant, State
 
-from .test_common import (
+from .common import (
     help_custom_config,
     help_test_availability_when_connection_lost,
     help_test_availability_without_topic,
